requirements.txt
pytest==7.3.1
faker==13.3.4
httpx~=0.25.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.3.1
pytest-xdist==3.3.1
python-dotenv==1.0.0
mysql-connector-python==8.0.33
requests==2.31.0
